Refactor, move AbstractMessageProcessor to different package
